COTECH-977 | Fix race condition
