+++ src/core/video/index.ts
@@ -1,4 +1,5 @@
 export * from './display-and-video-ads-sync-context';
+export * from './no-incontent-player-experiment';
 export * from './vast-debugger';
 export * from './vast-parser';
 export * from './video-event-provider';
